+++ union_sync_impl/src/main/java/com/example/union_sync_impl/entity/DocumentReserveCount.kt
@@ -1,10 +1,11 @@
 package com.example.union_sync_impl.entity
 
 import androidx.room.Entity
+import java.math.BigDecimal
 
 @Entity
 data class DocumentReserveCount(
     val id: String,
-    val count: Long?,
+    val count: BigDecimal?,
     val userUpdated: String?
 )
